llm_service/requirements.txt
# llm_service/requirements.txt
# Core dependencies for the API service
fastapi>=0.100.0 # FastAPI web framework
uvicorn>=0.20.0 # ASGI server for FastAPI
python-dotenv # For os.getenv, though env vars passed by compose

# Google Generative AI library
google-generativeai>=0.3.0 
